+++ view.go
@@ -114,11 +114,11 @@ func (f HavingLargerThanInt) Apply(event []Event) []Event {
 
 type HavingLargerThanFloat struct {
 	Name  string
-	Value float64
+	Value float32
 }
 
 func (f HavingLargerThanFloat) Apply(event []Event) []Event {
-	if event[len(event)-1].Record[f.Name].(float64) > f.Value {
+	if event[len(event)-1].Record[f.Name].(float32) > f.Value {
 		return event
 	}
 	return []Event{}
@@ -138,11 +138,11 @@ func (f HavingLessThanInt) Apply(event []Event) []Event {
 
 type HavingLessThanFloat struct {
 	Name  string
-	Value float64
+	Value float32
 }
 
 func (f HavingLessThanFloat) Apply(event []Event) []Event {
-	if event[len(event)-1].Record[f.Name].(float64) < f.Value {
+	if event[len(event)-1].Record[f.Name].(float32) < f.Value {
 		return event
 	}
 	return []Event{}
